Added utils classes for multinode Cassandra and startup independence ete tests to factor out code duplication. Removed unnecessary part of atlasStartsWithCassandraDownAndInitializesWithAllNodes
